+++ cogs/alerts.py
@@ -7,7 +7,7 @@ from discord.commands.context import ApplicationContext
 from discord.ext import commands
 
 import database
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.limiter import limit
 from util.logger import log
