+++ cogs/dictionary.py
@@ -10,7 +10,7 @@ from discord import option
 from discord.commands.context import ApplicationContext
 from discord.ext import commands
 
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.limiter import limit
 from util.logger import log
