+++ cogs/rules.py
@@ -2,7 +2,7 @@ import discord
 from discord.commands import option
 from discord.ext import commands
 
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.limiter import limit
 from util.logger import log
