+++ cogs/staffapps_backend.py
@@ -9,7 +9,7 @@ from discord import option
 from discord.ext import commands
 
 import database
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.limiter import limit
 from util.logger import log
