+++ cogs/staffapps_frontend.py
@@ -9,7 +9,7 @@ from discord.ext import commands
 from humanize import precisedelta
 
 import database
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.limiter import limit
 from util.logger import log
