+++ cogs/surveys.py
@@ -4,7 +4,7 @@ from os import getenv
 import discord
 from discord.ext import commands
 
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.logger import log
