+++ cogs/tips.py
@@ -2,7 +2,7 @@ from discord import Member, option
 from discord.commands.context import ApplicationContext
 from discord.ext import commands
 
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.limiter import limit
 from util.logger import log
