+++ cogs/wikipedia.py
@@ -3,7 +3,7 @@ import asyncio
 import wikipedia
 from discord.ext import commands
 
-from util.embed_builder import EmbedBuilder
+from message_formatting.embeds import EmbedBuilder
 from util.limiter import limit
 from util.logger import log
